backend/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile requirements.in
#
alembic==1.13.3
    # via flask-migrate
aniso8601==9.0.1
    # via flask-restx
annotated-types==0.7.0
    # via pydantic
anyio==4.6.2.post1
    # via
    #   httpx
    #   openai
attrs==24.2.0
    # via
    #   jsonschema
    #   referencing
beautifulsoup4==4.12.3
    # via -r requirements.in
blinker==1.8.2
    # via flask
boto3==1.35.54
    # via -r requirements.in
botocore==1.35.54
    # via
    #   boto3
    #   s3transfer
cachelib==0.9.0
    # via flask-caching
certifi==2024.8.30
    # via
    #   httpcore
    #   httpx
    #   requests
charset-normalizer==3.4.0
    # via requests
click==8.1.7
    # via flask
distro==1.9.0
    # via openai
flask==3.0.3
    # via
    #   -r requirements.in
    #   flask-caching
    #   flask-cors
    #   flask-migrate
    #   flask-redis
    #   flask-restx
    #   flask-sqlalchemy
flask-caching==2.3.0
    # via -r requirements.in
flask-cors==5.0.0
    # via -r requirements.in
flask-migrate==4.0.7
    # via -r requirements.in
flask-redis==0.4.0
    # via -r requirements.in
flask-restx==1.3.0
    # via -r requirements.in
flask-sqlalchemy==3.1.1
    # via
    #   -r requirements.in
    #   flask-migrate
greenlet==3.1.1
    # via sqlalchemy
h11==0.14.0
    # via httpcore
httpcore==1.0.6
    # via httpx
httpx==0.27.2
    # via openai
idna==3.10
    # via
    #   anyio
    #   httpx
    #   requests
importlib-resources==6.4.5
    # via flask-restx
itsdangerous==2.2.0
    # via flask
jinja2==3.1.4
    # via flask
jiter==0.7.0
    # via openai
jmespath==1.0.1
    # via
    #   boto3
    #   botocore
jsonschema==4.23.0
    # via flask-restx
jsonschema-specifications==2024.10.1
    # via jsonschema
mako==1.3.6
    # via alembic
markupsafe==3.0.2
    # via
    #   jinja2
    #   mako
    #   werkzeug
openai==1.53.0
    # via -r requirements.in
psycopg2==2.9.10
    # via -r requirements.in
psycopg2-binary==2.9.10
    # via -r requirements.in
pydantic==2.9.2
    # via openai
pydantic-core==2.23.4
    # via pydantic
python-dateutil==2.9.0.post0
    # via botocore
python-dotenv==1.0.1
    # via -r requirements.in
pytz==2024.2
    # via flask-restx
redis==5.2.0
    # via
    #   -r requirements.in
    #   flask-redis
referencing==0.35.1
    # via
    #   jsonschema
    #   jsonschema-specifications
requests==2.32.3
    # via -r requirements.in
rpds-py==0.20.1
    # via
    #   jsonschema
    #   referencing
s3transfer==0.10.3
    # via boto3
six==1.16.0
    # via python-dateutil
sniffio==1.3.1
    # via
    #   anyio
    #   httpx
    #   openai
soupsieve==2.6
    # via beautifulsoup4
sqlalchemy==2.0.36
    # via
    #   alembic
    #   flask-sqlalchemy
tqdm==4.66.6
    # via openai
typing-extensions==4.12.2
    # via
    #   alembic
    #   openai
    #   pydantic
    #   pydantic-core
    #   sqlalchemy
urllib3==2.2.3
    # via
    #   botocore
    #   requests
werkzeug==3.1.1
    # via
    #   flask
    #   flask-restx
